Move ResourceController and DownloadsController from Valhalla
