+++ app/controllers/scanned_resources_controller.rb
@@ -49,7 +49,7 @@ class ScannedResourcesController < BaseResourceController
       change_set.validate(file_metadata: [pdf_file])
       buffered_changeset_persister.save(change_set: change_set)
     end
-    redirect_to valhalla.download_path(resource_id: change_set.id, id: pdf_file.id)
+    redirect_to download_path(resource_id: change_set.id, id: pdf_file.id)
   end
 
   def save_and_ingest
