+++ app/controllers/templates_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 class TemplatesController < ApplicationController
-  include Valhalla::ResourceController
+  include ResourceController
   self.change_set_class = TemplateChangeSet
   self.resource_class = Template
   self.change_set_persister = ::ChangeSetPersister.new(
