+++ app/helpers/thumbnail_helper.rb
@@ -51,7 +51,7 @@ module ThumbnailHelper
     thumbnail_id = document.thumbnail_files.first.try :id
     return default_path unless thumbnail_id
 
-    url = valhalla.download_path document.id, thumbnail_id
+    url = download_path document.id, thumbnail_id
     return default_path unless url.present?
 
     image_tag url, image_options.merge(onerror: default_icon_fallback)
