+++ spec/models/ability_spec.rb
@@ -399,13 +399,13 @@ describe Ability do
     let(:vector_resource_members) { query_service.find_members(resource: vector_resource) }
     let(:file_set) { vector_resource_members.first }
     let(:vector_file) do
-      Valhalla::DownloadsController::FileWithMetadata.new(id: "1234", file: "", mime_type: "application/vnd.geo+json", original_name: "file.geosjon", file_set_id: file_set.id)
+      DownloadsController::FileWithMetadata.new(id: "1234", file: "", mime_type: "application/vnd.geo+json", original_name: "file.geosjon", file_set_id: file_set.id)
     end
     let(:metadata_file) do
-      Valhalla::DownloadsController::FileWithMetadata.new(id: "5768", file: "", mime_type: "application/xml; schema=fgdc", original_name: "fgdc.xml", file_set_id: file_set.id)
+      DownloadsController::FileWithMetadata.new(id: "5768", file: "", mime_type: "application/xml; schema=fgdc", original_name: "fgdc.xml", file_set_id: file_set.id)
     end
     let(:thumbnail_file) do
-      Valhalla::DownloadsController::FileWithMetadata.new(id: "9abc", file: "", mime_type: "image/png", original_name: "thumbnail.png", file_set_id: file_set.id)
+      DownloadsController::FileWithMetadata.new(id: "9abc", file: "", mime_type: "image/png", original_name: "thumbnail.png", file_set_id: file_set.id)
     end
 
     it {
