--- valhalla/app/controllers/valhalla/application_controller.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-module Valhalla
-  class ApplicationController < ActionController::Base
-    protect_from_forgery with: :exception
-  end
-end
